requirements.txt
glfw==2.7.0
numpy==1.26.4
pillow==10.3.0
PyOpenGL==3.1.7
